cdk/requirements.txt
aws-cdk-lib>=2.120.0,<3.0.0
constructs>=10.3.0,<11.0.0
python-dotenv>=1.0.0
